// frontend/src/api/models/createMemoryGroup.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * FastAPI
 * OpenAPI spec version: 0.1.0
 */

export interface CreateMemoryGroup {
  /**
     * @minLength 1
     * @maxLength 100
     */
  memoryGroupName: string;
  /**
     * @minLength 2
     * @maxLength 2
     */
  memoryLanguage: string;
  novelId: string;
}