// frontend/src/api/models/memoryGroup.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * FastAPI
 * OpenAPI spec version: 0.1.0
 */

/**
 * Memory group schema
 */
export interface MemoryGroup {
  /** Stable identifier to use when referring to this memory group. */
  memoryGroupId: string;
  /** Human-readable name of the memory group. */
  memoryGroupName: string;
  /** Language in which memory content should be written. */
  memoryLanguage: string;
  /** Identifier of the novel to which this memory group belongs. */
  novelId: string;
}